backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database (PostgreSQL for production, SQLite for development)
# psycopg2-binary==2.9.9  # Uncomment when PostgreSQL is installed

# Task Queue
celery==5.3.4
redis==5.0.1

# Environment
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Cryptography (for token encryption)
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2